+++ src/reprepbuild/utils.py
@@ -22,7 +22,16 @@
 
 import os
 
-__all__ = ("write_depfile",)
+__all__ = ("write_depfile", "write_dyndep")
+
+
+def filter_local_files(paths):
+    local = set()
+    for path in paths:
+        path = os.path.normpath(os.path.relpath(path))
+        if not path.startswith(".."):
+            local.add(path)
+    return sorted(local)
 
 
 def write_depfile(fn_depfile, outputs, inputs):
@@ -33,15 +42,25 @@ def write_depfile(fn_depfile, outputs, inputs):
     It is assumed that the depfile is always specified as "depfile = $out.depfile"
     and that there is only one output file.
     """
-    # Filter the inputs
-    deps = set()
-    for ipath in inputs:
-        ipath = os.path.normpath(os.path.relpath(ipath))
-        if not ipath.startswith(".."):
-            deps.add(ipath)
-
     with open(fn_depfile, "w") as f:
         f.write(" ".join(outputs))
         f.write(": \\\n")
-        for ipath in deps:
+        for ipath in filter_local_files(inputs):
             f.write(f"    {ipath} \\\n")
+
+
+def write_dyndep(fn_dd, fn_out, imp_outputs, imp_inputs):
+    """Write a dynamic dependency file for ninja, for a single output."""
+    with open(fn_dd, "w") as f:
+        f.write("ninja_dyndep_version = 1\n")
+        f.write(f"build {fn_out}")
+        imp_outputs = filter_local_files(imp_outputs)
+        if len(imp_outputs) > 0:
+            f.write(" | ")
+            f.write(" ".join(imp_outputs))
+        f.write(": dyndep")
+        imp_inputs = filter_local_files(imp_inputs)
+        if len(imp_inputs) > 0:
+            f.write(" | ")
+            f.write(" ".join(imp_inputs))
+        f.write("\n")