+++ CHANGELOG.md
@@ -7,10 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
-
 ### Fixed
 
+- Fixed other minor dependency tracking issues
+- Fixed dependency issues for Latex files.
 - Assume `SOURCE_DATE_EPOCH=315532800` (Jan 1, 1980) for compatibility with ZIP files.
 - When writing a build.ninja file, change directory before importing a python script.